move test to jvm
